requirements.txt
# Core AI & LangChain
langchain==0.1.0
langchain-openai==0.0.5
openai==1.10.0

# Vector Store
faiss-cpu==1.7.4

# Data Processing
pandas==2.1.4
numpy==1.26.2

# UI
streamlit==1.30.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# Development
black==23.12.1
pytest==7.4.3
